% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyAvg.R
\name{dailyAvg}
\alias{dailyAvg}
\title{Find an average over X values}
\usage{
dailyAvg(vec, every, na.rm = FALSE)
}
\arguments{
\item{vec}{a vector of values}

\item{every}{the number to average}

\item{na.rm}{remove NAs?}
}
\value{
a vector
}
\description{
Find an average over X values
}
\author{
Mike Johnson
}
