% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdc_functions.R
\name{fdc}
\alias{fdc}
\title{Calculate Flow Duration Curve from Timeseries}
\usage{
fdc(df, col_name = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{col_name}{The name of the column containing the streamflow values}
}
\value{
a data.frame
}
\description{
Given a data.frame and column of timeseries data, generate new column of flow excedence proabilites
}
\author{
Mike Johnson
}
