% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdc_functions.R
\name{fdc_curve}
\alias{fdc_curve}
\title{Calculate Flow Duration Spline}
\usage{
fdc_curve(df, col_name = NULL, prob = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{col_name}{The name of the column containing the streamflow values}

\item{prob}{Default NULL. If not NULL, function retruns data.frame of values at input excedence probabilites}
}
\value{
a curve equation or a data.frame
}
\description{
Given a data.frame and column of excedence proabilites, generate new column of flow excedence proabilites
}
\author{
Mike Johnson
}
