% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdc_functions.R
\name{fdc_plot}
\alias{fdc_plot}
\title{Plot Flow Duration Curve}
\usage{
fdc_plot(df, col_name = NULL, prob = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{col_name}{The name of the column containing the streamflow values}

\item{prob}{Default NULL. Input exedence probabilites (0-1) will be highlighted on plot}
}
\value{
a plot
}
\description{
Given a data.frame (that has already been processed by \code{fdc}) and column of streamflow timeseries, generate a plot
}
\author{
Mike Johnson
}
