% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Summarize model vs observe flows}
\usage{
get_summary(df, mod.name = "nwm", obs.name = "nwis")
}
\arguments{
\item{df}{a data.frame with a minumium of thress columns eg modeled, observed, and season}

\item{mod.name}{column name of modeled data}

\item{obs.name}{column name of observed data}
}
\value{
a vector of statistics
}
\description{
This function sumarizes a data frame produced by grabData or one that has a modeled and observered timeseries and a column of seasons
}
\author{
Mike Johnson
}
