% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabData}
\alias{grabData}
\title{Grab Data for a NWIS station from the USGS and NWM}
\usage{
grabData(stationID = NULL, retro.path = NULL, units = "cms",
  filter = TRUE)
}
\arguments{
\item{stationID}{staion id}

\item{retro.path}{path to retro_nwm netcdf file}

\item{units}{default is cms, alternativly use cfs}

\item{filter}{if filter is TRUE use filtered USGS meta data}
}
\value{
a data.frame
}
\description{
Grab Data for a NWIS station from the USGS and NWM
}
\author{
Mike Johnson
}
