% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Map NWIS Gaeges with metadata popup}
\usage{
map(stationID, filter = TRUE)
}
\arguments{
\item{stationID}{a vector of stationID(s)}

\item{filter}{use filtered data?}
}
\value{
a leaflet map
}
\description{
Map NWIS Gaeges with metadata popup
}
\author{
Mike Johnson
}
